Fixed two bugs
bug 1:  If a user would click the create button for a members volunteering, with no values, a null object would be created, which would crash table queries and prevent the /advanced/:id view from being loaded.

bug 2: More of a changed feature, but hours can now be saved as floats insted of integers.