Created /admin/dtb, /db/:tableName
Introduced *dynamic tables* which draw themselves every time that they are called. This replaced the previous static tables. The data is also loaded durring rendertime, and is not passed as a url string. This means that, unlike the narrow static table data cap, the dynamic tables have virtually no data cap on what can be rendered.

* note: DTB -> Dynamic Table Bulider
* note: DT  -> Dynamic Table